revise api/booking delete method

+++ app.py
@@ -67,12 +67,15 @@ class booking_tb(db.Model):
 	date = db.Column(db.String(50), nullable = False, comment = '預約日期')
 	time = db.Column(db.String(50), nullable = False, comment = '預約時間') 
 	price = db.Column(db.String(50), nullable = False, comment = '景點價格')
+	userId= db.Column(db.Integer, db.ForeignKey('user_tb.id'), comment='使用者訂購編號') 
+
 	
-	def __init__ (self,attractionId,date,time,price):
+	def __init__ (self,attractionId,date,time,price,userId):
 		self.attractionId = attractionId
 		self.date = date
 		self.time = time
 		self.price = price
+		self.userId = userId
 
 # Pages
 @app.route("/load")
@@ -122,7 +125,6 @@ def get_one_attraction(id):
 
 @app.route("/api/attractions", methods = ["GET"]) 
 def get_attraction_list():
-
 	try:
 		page = request.args.get("page", 0, type = int)
 		k = request.args.get("keyword")		
@@ -163,19 +165,16 @@ def get_attraction_list():
 
 @app.route("/api/user", methods = ["GET"])
 def get_user_data():
-	email=session.get('email')
 
-	if email:
-		user=user_tb.query.filter_by(email=email).first()
-		
-		user_data = {}
-		user_data['id'] = user.id
-		user_data['name'] = user.name
-		user_data['email'] = user.email
-		db.session.commit()
-		return jsonify({'data':user_data})
+	if "email" in session:
+			return jsonify({"data":{
+				"id":session["id"],
+				"name":session["name"],
+				"email":session["email"]
+				}
+			})
 	else:
-		return jsonify({'data':None})
+		return jsonify({"data":None}) 
 
 @app.route("/api/user", methods = ["POST"])
 def create_new_user():
@@ -207,22 +206,24 @@ def signin_user():
 
 	try:
 		if email and password:
-			user=user_tb.query.filter_by(email=email,password=password).first()
-			if not user:
-				return jsonify({"error":True, "message":"帳號或密碼錯誤,註冊失敗"}),400
-			else:
-				signin_email=data['email']
-				signin_password=data['password']
-				session['email']=signin_email
+			sql = f"select * from user_tb where email='{email}' and password='{password}'"
+			result = db.engine.execute(sql)
+			for row in result:
+				# print(i)
+				session["id"]=row[0]
+				session["email"]=row[1]
+				session["name"]=row[2]
 				return jsonify({"ok":True})
 		else:
-			return jsonify({"error":True})
+			return jsonify({"error":True})					
 	except:
 			return jsonify({"error":True, "message":"伺服器內部錯誤"}),500
 		
 @app.route("/api/user", methods = ["DELETE"])
 def signout_user():
 	if 'email' in session:
+		session.pop("id")
+		session.pop("name")
 		session.pop("email")
 		return jsonify({"ok":True})
 
@@ -231,17 +232,19 @@ def signout_user():
 def get_booking_data():
 	if 'email' not in session:
 		return jsonify({"error":True, "message":"未登入系統"}),403
+	
 	if request.method == "POST":
 		try:
 			data=request.get_json()
-			print(data)	
 			attractionId=data.get("attractionId")
 			date = data.get("date")
 			time = data.get("time")
 			price = data.get("price")
+			userId=session['id']
+
 			if date and time and price:
-				new_booking=booking_tb(attractionId=attractionId,date=date,time=time,price=price)
-				print(new_booking)
+				new_booking=booking_tb(attractionId=attractionId,date=date,time=time,price=price,userId=userId)
+				# print(new_booking)
 				db.session.add(new_booking)
 				db.session.commit()
 				return jsonify({"ok":True})
@@ -251,30 +254,33 @@ def get_booking_data():
 			return jsonify({"error":True, "message":"伺服器內部錯誤"}),500
 	
 	if request.method == "GET":
+		userId=session["id"]
 		
-		booking="SELECT booking_tb.*,attraction_tb.name,attraction_tb.address,attraction_tb.images FROM booking_tb INNER JOIN attraction_tb ON booking_tb.attractionId = attraction_tb.id"		
-		
+		booking=f'''
+		SELECT booking_tb.date,booking_tb.time,booking_tb.price,
+		attraction_tb.id as attId,attraction_tb.name,attraction_tb.address,attraction_tb.images,booking_tb.id 
+		FROM booking_tb
+		INNER JOIN attraction_tb ON booking_tb.attractionId = attraction_tb.id 
+		WHERE booking_tb.userId='{userId}' ORDER BY booking_tb.id DESC 
+		'''	
 		booking_data = db.engine.execute(booking)
-		res = {"data":None}
-		for row in booking_data:	
-			res = {
-				"data":{
+		for row in booking_data:
+			data = {
 					"attraction":{
-						"id":row[1],
-						"name":row[5],
-						"address":row[6],
-						"image":row[7].split(";")[0]
+						"id":row[3],
+						"name":row[4],
+						"address":row[5],
+						"image":row[6].split(";")[0]
 						},
-				"date":row[2],
-				"time":row[3],
-				"price":row[4]
+				"date":row[0],
+				"time":row[1],
+				"price":row[2]
 				}
-			}
-		return jsonify(res)
-
+			return jsonify({"data":data})
+		return jsonify({"data":None})
 	
 	if request.method=="DELETE":
-		db.session.query(booking_tb).delete()
+		db.session.query(booking_tb).filter_by(userId=session['id']).delete()
 		db.session.commit()
 		return jsonify({"ok":True})
 	return jsonify({"error":True})
